pythonProject/requirements.txt
django==3.1.1
djangorestframework
requests~=2.26.0
django-rest-swagger
python-dateutil~=2.8.2